# API.Web V12 (Documentação - Swagger + Scalar)/API.Web/Data/DTOs/V1/BookDTO.cs
namespace API.Data.DTOs.V1;

public class BookDTO
{     
    public long Id { get; set; }
    public string Title { get; set; } = default!;
    public string Author { get; set; } = default!;
    public decimal Price { get; set; }
    public DateTime LaunchDate { get; set; }
}