# API.Web V12 (Documentação - Swagger + Scalar)/API.Web/Data/DTOs/V2/PersonDTO.cs
using System.Text.Json.Serialization;

namespace API.Data.DTOs.V2;

public class PersonDTO
{   
    public long Id { get; set; }
    public string FirstName { get; set; } = string.Empty;
    public string LastName { get; set; } = string.Empty;
    public string Address { get; set; } = string.Empty;
    public string Gender { get; set; } = string.Empty;
    public DateTime? BirthDay { get; set; }
}